refactor: useUtilsを廃止
