+++ src/components/card/UserCard.tsx
@@ -1,20 +1,18 @@
 import React from 'react'
 import { View, Text, Linking } from 'react-native'
 import { Avatar } from 'react-native-elements/dist/avatar/Avatar'
-import { useUtils } from '@/hooks/useUtils'
 import { COLORS } from '@/constants/Colors'
 import { Ionicons } from '@expo/vector-icons'
 import { User } from '@/types/Foursquare'
 import { other } from '@/styles/styles'
 import { formatTimestamp } from '@/service/dateFns'
+import { generateImageUrl } from '@/service/utilFns'
 
 type Props = {
   user: User
 }
 
 const UserCard = ({ user }: Props) => {
-  const { generateImageUrl } = useUtils()
-
   const OpenProvider = (provider: string) => {
     void Linking.openURL(provider)
   }
