+++ src/components/pages/SignInByFoursquare.tsx
@@ -6,11 +6,10 @@ import { useFoursquare } from '@/hooks/useFoursquare'
 import { useNavigation } from '@react-navigation/native'
 import { FOURSQUARE_ACCESS_TOKEN } from '@/constants/StorageKeys'
 import { setUserId, logEvent } from '@/hooks/useAnalytics'
-import { useUtils } from '@/hooks/useUtils'
 import storage from '@/service/reactNativeStorage'
+import { parseURLParams } from '@/service/utilFns'
 
 const SignInByFoursquare = () => {
-  const { parseURLParams } = useUtils()
   const { fetchAccessToken, fetchUser } = useFoursquare()
   const { CLIENT_ID, REDIRECT_URI } = config()
   const navigation = useNavigation()
