+++ src/service/utilFns.ts
@@ -34,3 +34,41 @@ export const convertAgendaObject = (checkins: Checkin[]): AgendaItemsMap<Checkin
 
   return agendaObject
 }
+
+/**
+ * 画像URLを生成
+ * @param prefix
+ * @param suffix
+ * @param size サイズ
+ * @returns {string} URL
+ */
+export const generateImageUrl = (
+  prefix: string,
+  suffix: string,
+  size: number | string = 'original'
+) => {
+  return `${prefix}${size}${suffix}`
+}
+
+/**
+ * 誰かとチェックインした場合、シャウトの末尾に付いてしまう「〇〇と一緒に」を取り除く
+ */
+export const removeShoutWith = (shout: string) => {
+  const regObj = RegExp(/((— |.).*(と一緒に))/g)
+  return shout.replace(regObj, '')
+}
+
+/**
+ * URLのパラメータから指定のキーの値を取得する
+ * @param rawUrl 対象のURL
+ * @param key 取得したいパラメータのキー
+ * @returns 取得した値
+ */
+export const parseURLParams = (rawUrl: string, key: string) => {
+  const url = new URL(rawUrl)
+  const params = url.searchParams
+
+  if (params.has(key)) {
+    return params.get(key)
+  }
+}
